Fix operation_note int sizes #2272

+++ libraries/chain/include/steem/chain/operation_notification.hpp
@@ -12,9 +12,9 @@ struct operation_notification
 
    transaction_id_type trx_id;
    uint32_t            block = 0;
-   int32_t             trx_in_block = 0;
-   uint16_t            op_in_trx = 0;
-   uint64_t            virtual_op = 0;
+   uint32_t            trx_in_block = 0;
+   uint32_t            op_in_trx = 0;
+   uint32_t            virtual_op = 0;
    const operation&    op;
 };
